ANNPLT-57
Modify the Topic class to add a method for retrieving all "pending" announcements for topics which the user is an approver for.
Add an indicator on the topic list page displaying the number of announcements pending approval for all topics which the user is an approver for.
Add a "slide down" box that is activated when the above link is clicked, allowing the user to see the details for all announcements pending approval as well as approve the announcement or go directly to the edit page.